requirements.txt
streamlit
qrcode
pillow
uuid
time